add render util to prevent issues in email template viewer
